requirements.txt
appnope==0.1.4
asttokens==3.0.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
DateTime==5.5
debugpy==1.8.11
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.1.0
fonttools==4.56.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python==2.1.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyzmq==26.2.0
requests==2.32.3
scikit-surprise==1.1.4
scipy==1.13.1
setuptools==75.8.0
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
zipp==3.21.0
zope.interface==7.2
paramiko
psycopg2-binary
sqlalchemy
pyyaml
prometheus_client
mlflow
